chore: cleanup fvm configuration
I guess fvm changed the way it does it's configuration at some point,
but rather than having .fvm/fmv_config.json checked in .fvmrc should be
checked in. The .fvm directory is used to store files and symlinks
related to the installed Flutter version and as such should not be
commited at all.